initial plumbing for -display gtk
This works on a libvirt level configuration level. But when it comes
time to actually run the machine it fails because a background (likely
root) process is trying to open a gtk window for the user, and that
fails miserably.

Need to figure out the best way to have libvirt open local windows on
behalf or users. Or maybe this is not a good idea at all and we should
just stick with VNC.